Change tensor indexing form int to long

+++ src/model/loss/LoadedDiffusionMap.py
@@ -17,14 +17,14 @@ class LoadedDiffusionMap(DiffusionMap):
     def __call__(self, x: Tensor, y: Tensor) -> Tensor:
         self.D = self.D.to(x.device)
         border = 0.1
-        xc = x.ceil().int()
-        xf = x.floor().int()
+        xc = x.ceil().long()
+        xf = x.floor().long()
         xp = x - xf
         wxf = torch.where(xp <= border, 1, torch.where(xp >= 1 - border, 0, 1 - (xp - border) / (1 - 2 * border)))
         wxc = 1 - wxf
 
-        yc = y.ceil().int()
-        yf = y.floor().int()
+        yc = y.ceil().long()
+        yf = y.floor().long()
         yp = y - yf
         wyf = torch.where(yp <= border, 1, torch.where(yp >= 1 - border, 0, 1 - (yp - border) / (1 - 2 * border)))
         wyc = 1 - wyf
